Implement HybridRecommender with online SGD and ε-greedy exploration
