+++ botify/botify/config.json
@@ -1,27 +1,27 @@
 {
-  "REDIS_TRACKS_HOST": "redis",
+  "REDIS_TRACKS_HOST": "localhost",
   "REDIS_TRACKS_PORT": 6379,
   "REDIS_TRACKS_DB": 0,
 
-  "REDIS_ARTIST_HOST": "redis",
+  "REDIS_ARTIST_HOST": "localhost",
   "REDIS_ARTIST_PORT": 6379,
   "REDIS_ARTIST_DB": 1,
 
-  "REDIS_LISTEN_HISTORY_HOST": "redis",
+  "REDIS_LISTEN_HISTORY_HOST": "localhost",
   "REDIS_LISTEN_HISTORY_PORT": 6379,
   "REDIS_LISTEN_HISTORY_DB": 2,
 
-  "REDIS_RECOMMENDATIONS_LFM_HOST": "redis",
+  "REDIS_RECOMMENDATIONS_LFM_HOST": "localhost",
   "REDIS_RECOMMENDATIONS_LFM_PORT": 6379,
   "REDIS_RECOMMENDATIONS_LFM_DB": 3,
   "RECOMMENDATIONS_LFM_FILE_PATH": "./data/lightfm_i2i.jsonl",
 
-  "REDIS_RECOMMENDATIONS_SASREC_HOST": "redis",
+  "REDIS_RECOMMENDATIONS_SASREC_HOST": "localhost",
   "REDIS_RECOMMENDATIONS_SASREC_PORT": 6379,
   "REDIS_RECOMMENDATIONS_SASREC_DB": 4,
   "RECOMMENDATIONS_SASREC_FILE_PATH": "./data/sasrec_i2i.jsonl",
 
-  "REDIS_RECOMMENDATIONS_HSTU_HOST": "redis",
+  "REDIS_RECOMMENDATIONS_HSTU_HOST": "localhost",
   "REDIS_RECOMMENDATIONS_HSTU_PORT": 6379,
   "REDIS_RECOMMENDATIONS_HSTU_DB": 5,
   "RECOMMENDATIONS_HSTU_FILE_PATH": "./data/hstu_recommendations.json",
