+++ report.md
@@ -1,24 +1,27 @@
 # Отчёт: Улучшение Botify через онлайн‑ML HybridRecommender
 
 ## i. Abstract
-В данном эксперименте я заменяю стандартное ранжирование на **гибридную онлайн‑ML модель (contextual bandit)**. Модель обучается в режиме реального времени на логах сервиса, предсказывая ожидаемое время прослушивания трека на основе текущего контекста пользователя. В качестве контрольной группы используется **SasRec‑I2I**. Гипотеза заключалась в том, что персонализированное переранжирование кандидатов с использованием механизмов exploration позволит значительно увеличить $mean\_session\_time$ за счёт лучшего подбора контента под меняющиеся интересы пользователя.
+В данном эксперименте стандартное ранжирование было заменено на **гибридную онлайн‑ML модель (contextual bandit)**. Модель обучается в режиме реального времени, адаптируя веса различных источников рекомендаций на основе обратной связи от пользователя ($track\_time$). В качестве контрольной группы используется **SasRec‑I2I**. Гипотеза заключалась в том, что динамическое взвешивание кандидатов из разных источников (HSTU и I2I) с использованием механизма **exploration** позволит значительно увеличить среднее время сессии за счёт баланса между персональными интересами и релевантностью текущему треку.
 
 ## ii. Детали
-Реализованный **HybridRecommender** выступает в роли реранкера над пулом кандидатов. Кандидаты формируются из трех источников:
-1.  **User-based**: персонализированные списки из HSTU.
-2.  **Item-to-item**: рекомендации на основе «якорей» из недавней истории прослушивания через LightFM-i2i.
-3.  **Exploration**: случайные треки для борьбы с «пузырём фильтров».
+Реализованный **HybridRecommender** работает как адаптивный агрегатор кандидатов. Основные этапы работы:
+1.  **Генерация кандидатов**: Пул формируется из двух ключевых источников — персонализированных списков **HSTU** и рекомендаций на основе текущего трека (**I2I**).
+2.  **Вектор признаков (Features)**: Для каждого кандидата строится вектор признаков, включающий скоры из источников (позиция в выдаче) и константный признак (байес).
+3.  **Линейная модель**: Предсказание релевантности строится как скалярное произведение вектора признаков и весов пользователя, хранящихся в Redis.
+4.  **Exploration**: Реализована **$\epsilon$-greedy стратегия** (с параметром $\epsilon=0.1$), позволяющая системе открывать для пользователя новые треки и избегать зацикливания.
+5.  **Онлайн-обучение (SGD)**: При получении времени прослушивания предыдущего трека модель обновляет веса пользователя в Redis по правилу градиентного спуска, минимизируя ошибку предсказания.
 
-Для предсказания используется **линейная модель с SGD** и механизмом **feature hashing** (2^18 признаков). Признаки включают в себя метаданные трека (жанр, артист, страна, год), агрегаты истории пользователя (среднее время, доля скипов) и кросс-фичи (совпадение жанра/муда истории и кандидата). Модель обучается «честно»: данные о рекомендации временно сохраняются в Redis и обновляются, когда пользователь возвращает обратную связь в виде времени прослушивания на следующем запросе. Высокая производительность (снижение latency) достигнута за счёт кэширования хэшированных векторов признаков.
-
-### Диаграмма процесса:
-`history (redis)` → `candidates (HSTU + LightFM-i2i + explore)` → `features (ctx+item, hashing)` → `score` → `choose` → `log` → `SGD update (next request)`
+### Процесс:
+`history (redis)` → `candidates (HSTU + I2I)` → `linear scoring (weights from redis)` → `epsilon-greedy selection` → `SGD update on feedback`.
 
 ## iii. Результаты A/B эксперимента
-Эксперимент проводился по схеме: **Control = SasRec‑I2I** (`Treatment.C`), **Treatment = HybridRecommender** (`Treatment.T1`). Симуляция была запущена на **2000 эпизодах** с фиксированным сидом 3133. Скрипт `analyze_ab.py` подтвердил статистическую значимость улучшений по всем ключевым метрикам. Особенно выделяется рост времени сессии на **231%** и сокращение задержки ответа сервера почти на **100%** за счёт оптимизации вычислений.
+Эксперимент проводился по схеме: **Control = SasRec‑I2I** (`Treatment.C`), **Treatment = HybridRecommender** (`Treatment.T1`). Симуляция была запущена на **30 000 эпизодов**. Скрипт `analyze_ab.py` подтвердил высокую статистическую значимость улучшений. Главным достижением стал рост среднего времени прослушивания на сессию на **72.2%**.
 
 | treatment | metric | effect_pct | lower_pct | upper_pct | control_mean | treatment_mean | significant |
 | :--- | :--- | :--- | :--- | :--- | :--- | :--- | :--- |
-| T1 | **mean_time_per_session** | **231.40** | 215.02 | 247.77 | 1.3276 | 4.3995 | **True** |
-| T1 | mean_tracks_per_session | 49.94 | 45.97 | 53.92 | 6.3206 | 9.4774 | True |
-| T1 | mean_request_latency | -99.92 | -100.99 | -98.86 | 5902.6078 | 4.4612 | True |
+| T1 | **mean_time_per_session** | **72.20** | 68.87 | 75.53 | 2.2627 | 3.8964 | **True** |
+| T1 | **time (total)** | **76.00** | 71.16 | 80.84 | 7.4141 | 13.0489 | **True** |
+| T1 | mean_tracks_per_session | 22.38 | 21.25 | 23.50 | 7.2587 | 8.8829 | True |
+| T1 | mean_request_latency | 3.01 | 2.36 | 3.67 | 1.4279 | 1.4709 | True |
+
+**Вывод:** Внедрение гибридной модели с онлайн-обучением весов позволило значительно повысить вовлеченность пользователей. Несмотря на небольшой рост задержки (3.01%), вызванный дополнительными вычислениями и запросами к Redis, качество рекомендаций выросло критически.
